+++ src/Jade/Jade.php
@@ -3,6 +3,7 @@
 namespace Jade;
 
 use Jade\Compiler\FilterHelper;
+use Jade\Compiler\CacheHelper;
 use Jade\Parser\ExtensionsHelper;
 
 /**
@@ -350,84 +351,6 @@ class Jade extends Keywords
     }
 
     /**
-     * Return a file path in the cache for a given name.
-     *
-     * @param string $name
-     *
-     * @return string
-     */
-    protected function getCachePath($name)
-    {
-        return str_replace('//', '/', $this->options['cache'] . '/' . $name) . '.php';
-    }
-
-    protected function hashPrint($input)
-    {
-        // Get the stronger hashing algorithm available to minimize collision risks
-        $algos = hash_algos();
-        $algo = $algos[0];
-        $number = 0;
-        foreach ($algos as $hashAlgorithm) {
-            if (strpos($hashAlgorithm, 'md') === 0) {
-                $hashNumber = substr($hashAlgorithm, 2);
-                if ($hashNumber > $number) {
-                    $number = $hashNumber;
-                    $algo = $hashAlgorithm;
-                }
-                continue;
-            }
-            if (strpos($hashAlgorithm, 'sha') === 0) {
-                $hashNumber = substr($hashAlgorithm, 3);
-                if ($hashNumber > $number) {
-                    $number = $hashNumber;
-                    $algo = $hashAlgorithm;
-                }
-                continue;
-            }
-        }
-
-        return rtrim(strtr(base64_encode(hash($algo, $input, true)), '+/', '-_'), '=');
-    }
-
-    /**
-     * Return true if the file or content is up to date in the cache folder,
-     * false else.
-     *
-     * @param string  $input file or pug code
-     * @param &string $path  to be filled
-     *
-     * @return bool
-     */
-    protected function isCacheUpToDate($input, &$path)
-    {
-        if (is_file($input)) {
-            $path = $this->getCachePath(
-                ($this->options['keepBaseName'] ? basename($input) : '') .
-                $this->hashPrint(realpath($input))
-            );
-
-            // Do not re-parse file if original is older
-            return $this->getOption('upToDateCheck') && file_exists($path) && filemtime($input) < filemtime($path);
-        }
-
-        $path = $this->getCachePath($this->hashPrint($input));
-
-        // Do not re-parse file if the same hash exists
-        return file_exists($path);
-    }
-
-    protected function getCacheDirectory()
-    {
-        $cacheFolder = $this->options['cache'];
-
-        if (!is_dir($cacheFolder)) {
-            throw new \ErrorException($cacheFolder . ': Cache directory seem\'s to not exists', 5);
-        }
-
-        return $cacheFolder;
-    }
-
-    /**
      * Get cached input/file a matching cache file exists.
      * Else, render the input, cache it in a file and return it.
      *
@@ -440,20 +363,9 @@ class Jade extends Keywords
      */
     public function cache($input)
     {
-        $cacheFolder = $this->getCacheDirectory();
+        $cache = new CacheHelper($this);
 
-        if ($this->isCacheUpToDate($input, $path)) {
-            return $path;
-        }
-
-        if (!is_writable($cacheFolder)) {
-            throw new \ErrorException(sprintf('Cache directory must be writable. "%s" is not.', $cacheFolder), 6);
-        }
-
-        $rendered = $this->compile($input);
-        file_put_contents($path, $rendered);
-
-        return $this->stream($rendered);
+        return $cache->cache($input);
     }
 
     /**
@@ -465,38 +377,9 @@ class Jade extends Keywords
      */
     public function cacheDirectory($directory)
     {
-        $success = 0;
-        $errors = 0;
-        $cacheFolder = $this->getCacheDirectory();
+        $cache = new CacheHelper($this);
 
-        $extensions = new ExtensionsHelper($this->getOption('extension'));
-
-        foreach (scandir($directory) as $object) {
-            if ($object === '.' || $object === '..') {
-                continue;
-            }
-            $input = $directory . DIRECTORY_SEPARATOR . $object;
-            if (is_dir($input)) {
-                list($subSuccess, $subErrors) = $this->cacheDirectory($input);
-                $success += $subSuccess;
-                $errors += $subErrors;
-                continue;
-            }
-            if ($extensions->hasValidTemplateExtension($object)) {
-                $this->isCacheUpToDate($input, $path);
-                try {
-                    $contents = $this->compile($input);
-                    if (!file_put_contents($path, $contents) && strlen($contents)) {
-                        throw new \Exception('Write error', 1);
-                    }
-                    $success++;
-                } catch (\Exception $e) {
-                    $errors++;
-                }
-            }
-        }
-
-        return array($success, $errors);
+        return $cache->cacheDirectory($directory);
     }
 
     /**
